requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
Django==3.0.8
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-yasg==1.17.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
packaging==20.4
pycparser==2.20
pyleapcard==0.1.8
PyMySQL==0.10.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.10
